user delete his own replies, comments, reviews

+++ Gamezone/GameZone.Application/Comments/Commands/DeleteComment/DeleteCommentCommandHandler.cs
@@ -1,25 +1,34 @@
 ﻿using GameZone.Application.Interfaces;
+using GameZone.Domain.Models;
 using MediatR;
+using Microsoft.AspNetCore.Identity;
 
 namespace GameZone.Application.Comments.Commands.DeleteComment
 {
-    public class DeleteCommentCommandHandler : IRequestHandler<DeleteCommentCommand, Guid>
+    public class DeleteCommentCommandHandler : IRequestHandler<DeleteCommentCommand, string>
     {
         private readonly IUnitOfWork _unitOfWork;
+        private readonly UserManager<User> _userManager;
 
-        public DeleteCommentCommandHandler(IUnitOfWork unitOfWork)
+        public DeleteCommentCommandHandler(IUnitOfWork unitOfWork, UserManager<User> userManager)
         {
             _unitOfWork = unitOfWork;
+            _userManager=userManager;
         }
 
-        public async Task<Guid> Handle(DeleteCommentCommand request, CancellationToken cancellationToken)
+        public async Task<string> Handle(DeleteCommentCommand request, CancellationToken cancellationToken)
         {
             var comment = await _unitOfWork.CommentRepository.ReturnByIdAsync(request.Id);
+            var user = await _unitOfWork.UserRepository.ReturnByIdAsync(request.UserId);
 
-            await _unitOfWork.CommentRepository.DeleteAsync(comment);
-            await _unitOfWork.SaveAsync();
+            if (((request.UserId == comment.UserId) == true) || ((await _userManager.IsInRoleAsync(user, "Admin")) == true))
+            {
+                await _unitOfWork.CommentRepository.DeleteAsync(comment);
+                await _unitOfWork.SaveAsync();
+                return "Comment deleted";
+            }
 
-            return request.Id;
+            return "Not authorized";
         }
     }
 }
